Wrapperd the home component in an <IonPage>.
Without wrapping the content in <IonPage> the display for the <IonFab> component is broken, as it is pushed by the <IonHeader> component  and is only partly visible and mostly off screen. By adding the <IonPage> wrapper this problem is fixed and so the tutorial should be changed to accommodate it.

I fixed the code sinppets in the relevant parts of the guide.