WIP: add some documentation on IonicSlides
WIP: would be great to have  some documentation on IonicSlides actually do. Migrating from ion-slides I got the regression that pinch zoom on mobiles work, but doubleclick to zoom (desktop) stopped working. Puzzled because https://swiperjs.com/types/interfaces/types_modules_zoom.zoomoptions says toggle is default true - and then I came to think about IonicSlides.

Not sure this is the best place to put it though, but it would be convenient with a list of what settings IonicSlides overrides somewhere, and link to it.