docs(tabs): vue usage includes router outlet (#2685)
* Vue ion-router-outlet is missing, resulting in error

this isn't an edit about wrong stuff or not, but it's not clear why is it missing the ion-router-outlet while if I added it as-is it warns me of the lack of an ion router outlet. and it's confusing for why it's not referenced here. in addition to that it's not clear where should I add the tabs component and the schedule component etc

* docs(tabs): add router outlet imports to vue example