docs(react): clarify IonTabs behavior with IonPage
As noted in https://github.com/ionic-team/ionic-framework/issues/27705, it turns out that IonTabs in Ionic React automatically renders IonPage for you, so adding duplicate IonPage instances can cause issues. This behavior is not aligned with Vue, so we should look into making this consistent in the future. For now, I've updated the docs to make this clear.