Update slides.md
on ionic 7.5.5, swiper 11.0.4 the table above should probably be rewritten to have the prefix "swiper" to all the events. 

event called "init" in table -> swiperinit
and so on for all events.


There are forum answers putting us on this track
https://forum.ionicframework.com/t/ionic-7-angular-and-swiper-11-0-3-event-did-not-trigger/237358

Nice caveat I found, did trick me at first:

The (click) event DOES work without the "swipe" prefix, but that's a generic click handler, not the Swipe click handler:

<swiper-container (click)="handler()"
receives a regular click event

<swiper-container (swiperclick)="handler()"
event.detail contains pointerup data. This is the ACTUAL Swipe (click) implementation, see https://swiperjs.com/swiper-api#event-click


Didn't edit the full table -- I'm not a guru and I could be barking up the wrong tree here, so I added it as a note for now    o_O